Feat: 네트워크가 없을떄에 list불러오기 기능 추가

+++ app/src/main/java/com/myboard/data/repository/ContentRepositoryImpl.kt
@@ -7,6 +7,7 @@ import com.myboard.data.source.local.dao.ContentDao
 import com.myboard.data.source.remote.api.ContentService
 import com.myboard.domain.model.Content
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.flow
 import java.io.IOException
 import javax.inject.Inject
@@ -17,8 +18,14 @@ class ContentRepositoryImpl @Inject constructor(
 ) : ContentRepository {
     override suspend fun save(item: Content): Boolean {
         return try {
-            contentService.saveItem(item.toRequest())
-            contentDao.insert(item.toEntity())
+            contentService.saveItem(item.toRequest()).also {
+                if (it.success) {
+                    //서버에 저장이 완료되었을때
+                    it.data?.let { contentDto ->
+                        contentDao.insert(contentDto.toEntity())
+                    }
+                }
+            }
             true
         } catch (e: IOException) {
             false
@@ -27,8 +34,13 @@ class ContentRepositoryImpl @Inject constructor(
 
     override suspend fun update(item: Content): Boolean {
         return try {
-            contentService.updateItem(item.toRequest())
-            contentDao.insert(item.toEntity())
+            contentService.updateItem(item.toRequest()).also {
+                if (it.success) {
+                    it.data?.let { contentDto ->
+                        contentDao.insert(contentDto.toEntity())
+                    }
+                }
+            }
             true
         } catch (e: IOException) {
             false
@@ -38,9 +50,12 @@ class ContentRepositoryImpl @Inject constructor(
     override suspend fun delete(item: Content): Boolean {
         return try {
             item.id?.let { id ->
-                contentService.deleteItem(id)
+                contentService.deleteItem(id).also {
+                    if (it.success) {
+                        contentDao.delete(item.toEntity())
+                    }
+                }
             }
-            contentDao.delete(item.toEntity())
             true
         } catch (e: IOException) {
             false
@@ -49,18 +64,18 @@ class ContentRepositoryImpl @Inject constructor(
 
     override fun loadList(): Flow<List<Content>> {
         return flow {
-//            contentDao.selectAll().collect{list ->
-//                emit(//room db 조회
-//                    list.map { it.toContent() }
-//                )
-//            }
-            emit(//api 통신을 통한 디비 조회
-                try {
-                    contentService.getList().data.map { it.toContent() }
-                } catch (e: IOException) {
-                    emptyList()
+            //네트워크가 없을때에도 작동하기 위함
+            try {
+                contentService.getList().data.also { list ->
+                    contentDao.insertAll(list.map { it.toEntity() })
                 }
-            )
+            } finally {
+                contentDao.selectAll().collect { list ->
+                    emit(
+                        list.map { it.toContent() }
+                    )
+                }
+            }
         }
     }
 }
